the rest of the long longs

+++ cuda_functions/nms_3D/src/cuda/nms_kernel.cu
@@ -28,7 +28,7 @@ __device__ inline float devIoU(float const * const a, float const * const b) {
 }
 
 __global__ void nms_kernel(const int n_boxes, const float nms_overlap_thresh,
-                           const float *dev_boxes, unsigned long long *dev_mask) {
+                           const float *dev_boxes, int64_t *dev_mask) {
   const int row_start = blockIdx.y;
   const int col_start = blockIdx.x;
 
@@ -62,7 +62,7 @@ __global__ void nms_kernel(const int n_boxes, const float nms_overlap_thresh,
     const int cur_box_idx = threadsPerBlock * row_start + threadIdx.x;
     const float *cur_box = dev_boxes + cur_box_idx * 7;
     int i = 0;
-    unsigned long long t = 0;
+    int64_t t = 0;
     int start = 0;
     if (row_start == col_start) {
       start = threadIdx.x + 1;
@@ -79,7 +79,7 @@ __global__ void nms_kernel(const int n_boxes, const float nms_overlap_thresh,
 
 
 void _nms(int boxes_num, float * boxes_dev,
-          unsigned long long * mask_dev, float nms_overlap_thresh) {
+          int64_t * mask_dev, float nms_overlap_thresh) {
 
 
   dim3 blocks(DIVUP(boxes_num, threadsPerBlock),
